Add files via upload
The car sales data is analyzed using the SQL server management tool and is populated into the database with the help of Visual Studio and SQL server integration services (SSIS). These are then populated to the database. With this regard, car sales information is taken to understand and implement better decisions.


Online customer engagement and sales
 Dealership management.
 Network consolidation to match demand.
 Reduced profits for parts servicing electric vehicles with fewer moving parts.
 New types of finance and transportation-as-a-service.


 To acquire essential business metrics rapidly to facilitate accessibility for business data to all the employees in the company.
- Removes the need for speculation – Dealership reputation is always at stake.
- Valuable insights into customer behavior through data analysis and identification of risk.